bug fixes - If the RADDOSE-3D input file already existed in the main directory with the same name as the self.RADDOSEfilename string then an error was thrown. I have now added a check to see if this is the case.
- The code block that checks the input file for possible errors was placed
after the code that runs RADDOSE-3D. I've now moved this before running
RADDOSE-3D so the checks are done first.

- I've added try: statement over the code that parses the RADDOSE-3D input
file. This is because if the user was missing an essential line in the input
file then the parser would throw an error and wouldn't tell the user. This
is a temporary fix. Really the parser should also check that essential lines
are present and let the user know if it isn't.